% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tab.R
\name{Abb}
\alias{Abb}
\title{Abbildungs- Nummer}
\usage{
Abb(filename = "", caption = "", folder = options()$stp25$fig_folder)
}
\arguments{
\item{filename, caption}{Name der Grafik und Titel}

\item{folder}{Speicherort}
}
\value{
Name plus Pfad
}
\description{
Nummerierte Abbildung default Folder ist fig
}
\author{
Wolfgang Peter
}
