% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_library.R
\name{citation_library}
\alias{citation_library}
\title{citation_library}
\usage{
citation_library(..., style = "text", .bibstyle = NULL, output = "text")
}
\arguments{
\item{...}{namen der Librarys}

\item{style, .bibstyle}{an format}

\item{output}{html oder text}
}
\value{
character
}
\description{
citation_library
}
\examples{

citation_library(base,Hmisc,car,lattice)

}
