% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.data.frame.R
\name{cleanup_nbsp}
\alias{cleanup_nbsp}
\title{nbsp entfernen}
\usage{
cleanup_nbsp(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
Interne Funktion Html &nbsp; erntfernen
}
