% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\alias{Sprachdatei}
\alias{Names2Language}
\title{dictionary (Sprachdatei)}
\usage{
Sprachdatei(language = "de", words = list(characteristics =
  c("Characteristics", "Bezeichnung"), statistics = c("Statistics",
  "Statistik"), statistic = c("Statistics", "Statistik"), mean.Rank =
  c("mean Rank", "mittlerer Rang"), skalenwert = c("Characteristic Value",
  "MerkmalsausprAegung"), items = c("Items", "Items"), item = c("Item", "Item"),
  sig.test = c("significance Test", "signifikanz Test"), value = c("Value",
  "Wert"), variable = c("Variable", "Merkmal"), mass = c("Value",
  "Anpassungsmass"), annforderung = c("Requirement", "Anforderung"),      source
  = c("Source", "Quelle"), N = c("N", "N"), n = c("n", "n"), mean = c("Mean",
  "M"), sd = c("SD", "SD"), ci95.low = c("low 95\% CI", "low 95\% CI"), ci95.hig
  = c("up 95\% CI", "up 95\% CI"), se = c("SE", "SE"), estimate = c("B", "B"),
  est.std = c("Beta", "Beta"), beta = c("Beta", "Beta"), df = c("Df", "Df"),
  z.value = c("z", "z"), f.value = c("F", "F"), t.value = c("T", "T"), p.value =
  c("p", "p"), lr.chisq = c("LR Chisq", "LR Chisq"), eta = c("eta Sq",
  "eta Sq"), adj.r2 = c("adj. R2", "adj. R2"),      sum.sq = c("Sum Sq",
  "Sum Sq"), r2 = c("R2", "R2")))

Names2Language(words)
}
\arguments{
\item{language}{Sprache default="de"}

\item{words}{Wort-liste Deutsch Englisch}
}
\description{
Funktionen die Texte fuer die Beschriftungen bereitstellen stp25

Liste mit Uebersaetzung.

Names2Language() wird von Output benutzt um die De/En Uebersaetzun
 durchzufuehren.
}
\examples{
Names2Language(c("Pr..Chisq.", "F.value"))
}
