% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_format.R, R/fix_irgendwas.R
\name{fix_format}
\alias{fix_format}
\alias{fix_irgendwas}
\alias{fix_format_p}
\alias{fix_colnames}
\alias{fix_data_frame2}
\alias{fix_levels}
\title{Fix Objekt zu Data.Frame}
\usage{
fix_format(x, exclude = NULL, digits = NULL,
  pattern_pval = c("Pr\\\\(\\\\>", "Pr\\\\.\\\\.", "p\\\\.value", "p value",
  "pvalue"), pattern_est = c("Estimate", "Std\\\\. Error", "est.std", "se"),
  pattern_df = c("Df"), pattern_N = c("N", "n"),
  apa.style = options()$stp25$apa.style, ...)

fix_format_p(x, df1 = NULL, df2 = null, p = NULL)

fix_colnames(x, col_names = NULL, translate = TRUE)

fix_data_frame2(...)

fix_levels(x, labels = NULL, which = ncol(x) - 1,
  measure.vars = unique(x[, which]), ...)
}
\arguments{
\item{x}{vektor liste oder matrix}

\item{exclude}{Position der auszuschliedenden Spalten}

\item{digits}{exakt nach muster Runden}

\item{pattern_pval}{fuer fix_format}

\item{pattern_est}{fuer fix_format}

\item{pattern_df}{fuer fix_format}

\item{pattern_N}{N Muster}

\item{apa.style}{fuer fix_format}

\item{...}{alles weiter an data.frame}

\item{df1, df2, }{Dichte}

\item{p}{P-Wert}

\item{col_names}{fuer fix_colnames}

\item{translate}{fuer fix_colnames}

\item{labels}{fix_levels labels als String oder data.frame}

\item{which}{fix_levels Indexnummer der Variable ncol(x)-1}

\item{measure.vars}{fix_levels welche Variablen}
}
\value{
gleiches Objekt
}
\description{
Repariert Levels und labels und Zahlen

fix_format_p sucht automatisch nach den p-Werten die meist an der Letzten stelle sind
und gibt einen Vector-String mit der LAenge drei aus.
Nicht zu verwechseln mit \code{rndr_P()}
}
\details{
fix_to_data_frame input Matrix over Vektor Output( Source =Colnames + Daten )
fix_format_p Input(F, df,und p)  Output: (test, df, p.value)

fix_levels ist fuer dta %>% gather(Parameter, Wert, x1:x5)  %>% fix_levels(dta)
fix_colnames ist fuer Formatieren in der Funktion \code{Output()}
}
\examples{
  df <- data.frame(Item=c("a", "b"),
x=1:2,
x2=c(1.2,2.3),
beta=  c(.22,.13),
est=c(2.4234,.03),
p.value=c(0.02456,0.0398))

# Format2(df , digits=c( NA,0,1,2,3,4))

#fix_format(df,  exclude=2)
#fix_format(df,  digits=c( NA,0,1,2,3,4))
}
