% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25_fix_to_data_frame.R
\name{fix_to_data_frame}
\alias{fix_to_data_frame}
\alias{fix_to_data_frame.default}
\alias{fix_to_data_frame.ftable}
\alias{fix_to_data_frame.table}
\title{Transformiere zu data.frame
fix_format_p Input(F, df,und p)  Output: (test, df, p.value)}
\usage{
fix_to_data_frame(x, ...)

\method{fix_to_data_frame}{default}(x, ...)

\method{fix_to_data_frame}{ftable}(x, ncol = length(attributes(x)$col.vars),
  nrow = length(attributes(x)$row.vars), nlength = ncol + nrow, ...)

\method{fix_to_data_frame}{table}(x, ...)
}
\arguments{
\item{x}{vektor liste oder matrix}

\item{...}{weiter Argumente}

\item{ncol}{nicht zum aendern}

\item{nrow}{nicht zum aendern}

\item{nlength}{nicht zum aendern}
}
\value{
ein data.frame Objekt
}
\description{
fix_levels ist fuer dta %>% gather(Parameter, Wert, x1:x5)  %>% fix_levels(dta)
fix_colnames ist fuer Formatieren in der Funktion \code{Output()}
}
