% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projekt.R
\name{get_scriptpath}
\alias{get_scriptpath}
\title{Ausfuehrendes File finden}
\usage{
get_scriptpath()
}
\value{
String
}
\description{
In Projekt verwendet um den Namen des Files zu dokumentieren https://stackoverflow.com/questions/18000708/find-location-of-current-r-file
}
