% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML.R
\name{HTML_}
\alias{HTML_}
\alias{HTML_P}
\alias{HTML_I}
\alias{HTML_B}
\alias{HTML_S}
\alias{HTML_HR}
\alias{HTML_DIF}
\alias{HTML_CENTER}
\alias{HTML_NULL}
\alias{HTML_BR}
\title{HTML Text Formatting Elements}
\usage{
HTML_(x, output = which_output())

HTML_P(x, output = which_output())

HTML_I(x, output = which_output())

HTML_B(x, output = which_output())

HTML_S(x, output = which_output())

HTML_HR(x, output = which_output())

HTML_DIF(x, output = which_output())

HTML_CENTER(x, output = which_output())

HTML_NULL(x, output = which_output())

HTML_BR(x, output = which_output())
}
\arguments{
\item{x}{Text}

\item{output}{HTML, Knit  oder Textfile (ist nicht gedacht zum Aendern)}
}
\value{
Null doere Character-String
}
\description{
HTML_ schreibt direkt ins HTML-File

HTML_P <p> Paragraph

HTML_I <i>	Defines italic text

HTML_B <b>	Defines bold text

HTML_S <small>	Defines smaller text

HTML_HR <hr> Lienie

HTML_DIF <dif> Formatierungsbereich

HTML_CENTER <center> zentrieren

HTML_NULL wie  HTML_BR <br> Zeilenumbruch aber mit Text

HTML_BR <br> Zeilenumbruch aber ohne Text
}
\examples{
# HTML_P( "Hallo Welt", file="")

}
