% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myset.R
\name{MySet}
\alias{MySet}
\title{MySet}
\usage{
MySet(col = NULL, pch = 15:18, lty = 1:3, cex = 1,
  col.bar = NULL, n = if (is.numeric(col)) col else 4, h.start = 120,
  theme = latticeExtra::ggplot2like(n = n, h.start = h.start),
  axis.grid = FALSE, knit = stp25output::which_output() == "markdown",
  ...)
}
\arguments{
\item{col}{Farben brewer.pal(9,"Set1")}

\item{pch}{Symbole}

\item{lty}{Striche und Lienien}

\item{cex}{Punkt in Boxplots}

\item{col.bar}{Farbe von einzelnen Balken}

\item{n}{anzahl an Farben default ist 4}

\item{h.start}{Farben = 120,}

\item{theme}{LaticeExtra ggplot2like(), #theEconomist.theme() custom.theme()}

\item{axis.grid}{Referenzlinien = FALSE,}

\item{knit}{einstellung für knitter}

\item{...}{nicht verwendet}
}
\value{
Gibt opar zureuck (Wird fue reset benötigt)
}
\description{
Initialisiert Lattice ladet die nötigen Pakete und Formatiert die Lattice Grafiken
Einstellen der Farben und Symbole ueber \link{trellis.par.set}
}
\examples{
#graphics.off()
 
 
   lattice::trellis.par.set(effects::effectsTheme())
   
 
 
 
 
#library(RColorBrewer)
# Set3 rosa-himmelblau
# brewer.pal(8,"Set3")[c(3,4)]


 #- display.brewer.pal(n = 4, name = 'Set2')
 #-
 #- brewer.pal.info
 #- Diverging palettes
 #-          maxcolors category colorblind
 #- BrBG            11      div       TRUE  braun-gruen
 #- PiYG            11      div       TRUE  violet-gruen
 #- PRGn            11      div       TRUE  violet-gruen
 #- PuOr            11      div       TRUE  braun-violett
 #- RdBu            11      div       TRUE  rot-blau
 #- RdGy            11      div      FALSE  rot-grau
 #- RdYlBu          11      div       TRUE  rot-blau
 #- RdYlGn          11      div      FALSE  rot-gruen
 #- Spectral        11      div      FALSE  rot-blau
 #- Qualitative palettes
 #-          maxcolors category colorblind
 #- Accent           8     qual      FALSE
 #- Dark2            8     qual       TRUE
 #- Paired          12     qual       TRUE
 #- Pastel1          9     qual      FALSE
 #- Pastel2          8     qual      FALSE
 #- Set1             9     qual      FALSE
 #- Set2             8     qual       TRUE
 #- Set3            12     qual      FALSE
 #- Sequential palettes
 #-          maxcolors category colorblind
 #- Blues            9      seq       TRUE
 #- BuGn             9      seq       TRUE
 #- BuPu             9      seq       TRUE
 #- GnBu             9      seq       TRUE
 #- Greens           9      seq       TRUE
 #- Greys            9      seq       TRUE
 #- Oranges          9      seq       TRUE
 #- OrRd             9      seq       TRUE
 #- PuBu             9      seq       TRUE
 #- PuBuGn           9      seq       TRUE
 #- PuRd             9      seq       TRUE
 #- Purples          9      seq       TRUE
 #- RdPu             9      seq       TRUE
 #- Reds             9      seq       TRUE
 #- YlGn             9      seq       TRUE
 #- YlGnBu           9      seq       TRUE
 #- YlOrBr           9      seq       TRUE
 #- YlOrRd           9      seq       TRUE

 #Projekt( )
 #MySet(brewer.pal(9,"Set1")[c(8,2)], col.bar=3 )
 #windows(8,8)
 #  show.settings()
 #windows(7,4)
 #  bwplot(yield ~ site|year, barley )
 #  End()
}
