% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_Ordnen.R
\name{Ordnen}
\alias{Ordnen}
\alias{Ordnen.default}
\alias{Ordnen.eff}
\alias{Ordnen.efflist}
\title{Ordnen (Tidy)}
\usage{
Ordnen(x, ...)

\method{Ordnen}{default}(x, ...)

\method{Ordnen}{eff}(x, ...)

\method{Ordnen}{efflist}(x, ...)
}
\arguments{
\item{x}{Objekt}

\item{...}{weitere Objekte nicht benutzt}
}
\value{
ein data.frame-Ojekt oder eine Liste von data.frames. Im Attribut N sind die Stichprobengroesse
und notes
}
\description{
Diese Funktion meine Version von tidy
}
