% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R, R/apa_table.R, R/data.frame.R,
%   R/kable.R, R/xtabs.R
\name{Output}
\alias{Output}
\alias{which_output}
\alias{Output.vector}
\alias{Output.htmlTable}
\alias{Output.NULL}
\alias{Output.default}
\alias{text_as_table}
\alias{Caption}
\alias{Output.apa_table}
\alias{Output.matrix}
\alias{Output.list}
\alias{Output.stp25}
\alias{Output.data.frame}
\alias{Output_kable}
\alias{Output_kable.list}
\alias{Output_kable.default}
\alias{Output.table}
\alias{Output.xtabs}
\alias{Output.ftable}
\title{Ausgabe von HTML, Knit oder Text
 
Erstellt den Output von Objekten, diese koennen zB. mit Uberschriften versehen sein.}
\usage{
Output(x, ...)

which_output(...)

\method{Output}{vector}(x, output = options()$prompt[1] == "HTML> ", ...)

\method{Output}{htmlTable}(x, output = which_output())

\method{Output}{NULL}(x, ...)

\method{Output}{default}(x, caption = "", note = "",
  output = options()$prompt[1] == "HTML> ", col_names = NULL,
  print_col = NULL, ...)

text_as_table(x, ...)

Caption(x = NULL, atr = NULL, output = which_output())

\method{Output}{apa_table}(x)

\method{Output}{matrix}(x, ...)

\method{Output}{list}(x, ...)

\method{Output}{stp25}(x, ...)

\method{Output}{data.frame}(x, caption = NULL, note = NULL,
  output = which_output(), print_col = NULL, col_names = NULL,
  fix_colnames = TRUE,
  css.table = "padding-left: .5em; padding-right: .2em;",
  css.cell = "padding-left: .5em; padding-right: .2em;", ...)

Output_kable(x, ...)

\method{Output_kable}{list}(x, ...)

\method{Output_kable}{default}(x, caption = NULL, col.names = colnames(x),
  booktabs = TRUE, note = NULL, output = which_output(),
  print_col = NULL, fix_colnames = TRUE, ...)

\method{Output}{table}(x, caption = attr(x, "caption"), note = attr(x,
  "note"), ...)

\method{Output}{xtabs}(x, caption = attr(x, "caption"), note = attr(x,
  "note"), ...)

\method{Output}{ftable}(x, caption = attr(x, "caption"), note = attr(x,
  "note"), ...)
}
\arguments{
\item{x}{Objekt liste oder Dataframe}

\item{...}{weitere Einstellungen}

\item{output}{HTML, Knit  oder Textfile (ist nicht gedacht zum Aendern)}

\item{caption, note}{Ueberschrift Fussnote}

\item{col_names, fix_colnames}{Spalten Ueberschrift aendern (data.frame)}

\item{print_col}{Spalten nicht Ausgeben (data.frame)}

\item{atr}{in Caption: alternativer Text}

\item{css.table, css.cell}{Format an Output.data.frame (htmlTable)
padding-left: .5em; padding-right: .2em; Output.data.frame (htmlTable)}

\item{col.names}{Output_kable:  fuer tintPdf format = "latex"}

\item{booktabs}{kable an Latex Latex}
}
\value{
HTML oder Textausgabe
}
\description{
Ausgabe von HTML, Knit oder Text
 
Erstellt den Output von Objekten, diese koennen zB. mit Uberschriften versehen sein.

Output.character fuer brerits erstellten Html-Code

Output.htmlTable: Verarbeitet htmlTable-Objekt als auch Listen
 oder auch einfach nur ein string. Die Tabelle mit Header usw ist dabei
 schon vorbereitet. Bei Markdown kommt output=FALSE

Ueberschrift aus stp-Objekt: Caption(caption, attr(x, "caption"))
oder Note(note, attr(x, "note")) mit

Output fur apaTables-Objekte wie zB apa.reg.table

Output.matrix: umwandeln in einen data.frame

Output.iste: einzeln in einen data.frame transformieren

Output.stp25: experimenteller Prototyp

Output.data.frame ist die Standart-Funktion fuer die
Ausgabe. Sie Arbeitet mit \code{htmlTable}, die einzelnen
Header-Ebenen werden ueber Header1_M, Header1_SD gesteuert.
Auch die Funktionen Output.table und xtabl arbeiten ueber diese Funktion.

Output mit  knitr::kable Funftioniert nur wenn options(knitr.table.format = "latex")
gesetzt wird (das gilt nur fuer Pdf)

Output.xtabs und Output.tabl und Output.ftable fuer Tabellen Objekte
}
\examples{

\dontrun{
library(apaTables) 
 basic.reg <- lm(sales ~ adverts + airplay, data=album)
apa.reg.table(basic.reg) \%>\% Output()


lm_output <- lm(libido ~ dose, data=viagra)
apa.aov.table(lm_output) \%>\% Output()

APA_Table(lm_output)

APA_Table( aov(libido ~ dose, data=viagra))

block1 <- lm(sales ~ adverts + airplay, data=album)
block2 <- lm(sales ~ adverts*airplay, data=album)
apa.reg.table(block1, block2) \%>\% Output()
 } 

}
\author{
Wolfgang Peter
}
