% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_word.R
\name{Output_word}
\alias{Output_word}
\title{Output nach docx}
\usage{
Output_word(
  x,
  caption = NULL,
  note = NULL,
  output = which_output(),
  print_col = NULL,
  col_names = NULL,
  fix_colnames = options()$stp25$language != "",
  ...
)
}
\arguments{
\item{x}{data.frame}

\item{caption, note, output, print_col, col_names, fix_colnames}{Ueberschrift usw}

\item{...}{not used}
}
\value{
flextable
}
\description{
Geht nur direkt asis_output() function only works in top-level R expressions,
}
