% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projekt.R
\name{Projekt}
\alias{Projekt}
\alias{End}
\alias{Methode}
\alias{Materials}
\alias{Research_Design}
\alias{Measures}
\alias{Results}
\alias{Demographic_Variables}
\alias{Statistic}
\title{Projekt Funktion}
\usage{
Projekt(myformat = "", Projektname = "Demo", datum = date(),
  fig_folder = "Fig", html_folder = "Results", OutDec = NULL,
  contrasts = c("contr.Treatment", "contr.poly"),
  html_name = stpvers::Clean_Umlaute2(Projektname), ...)

End(anhang = FALSE,
  browser = "C:/Program Files (x86)/Internet Explorer/iexplore.exe",
  output = options()$prompt[1] == "HTML> ", ...)

Methode(x = NULL, file = NULL)

Materials(x = NULL, file = "(1) Get Data.R")

Research_Design(x = NULL, file = NULL)

Measures(x = NULL, file = "(2) Measures.R")

Results(x = NULL, file = NULL)

Demographic_Variables(x = NULL, file = "(3) Demographic.R")

Statistic(x = "Resultate", file = "(4) Analyse.R")
}
\arguments{
\item{myformat}{HTML oder Knit oder Console}

\item{Projektname}{Bezeichnung des Projektes (gilt auch fuer die HTML Seite)}

\item{datum}{Datum zur Dokumentation}

\item{fig_folder}{Folder wenn ein ander Ort gewuenscht}

\item{html_folder}{Folderwenn ein ander Ort gewuenscht}

\item{OutDec}{Komma oder Punkt}

\item{contrasts}{default wie SPSS}

\item{html_name}{intern nicht aendern}

\item{...}{weitere Objekte nicht benutzt}

\item{anhang}{Ja/Nein}

\item{browser}{Ie oder Chrome}

\item{output}{TRUE oder FALSE}

\item{x}{Character}

\item{file}{auszufuerendes File}
}
\value{
HTML oder Text Output
}
\description{
Adaption der R2HTML-Funktionen. Und Setzen von Default Einstellungen wie Contraste und set.seed(0815).

Bei ausführen von library(stp25output) werden die Optionen mit default_stp25_opt() gesetzt.

Projekt() Initialisiere ein neues Projekt, und erstellt dabei die Folder Results und Fig 
mit End() wird der Report aufgerufen und ein Reset der Einstellungen durchgeführt. Ueber die Funktionen
stp25_options werden Formatierungsoptionen gesetzt.
\subsection{Projekt und HTML}{
Die Funktion \code{Projekt}  ist essentiell fuer alle weiteren Funktionen es werden dabei
Variablen fuer die Darstellung gesetzt wie ZB Nachkommastellen
\code{digits.prozent = 0, digits.mean = 2, digits.r = 2,  digits.p = 3, digits.Fstat = 2}
Trennzeichen \code{OutDec = "."}. Das Aussehen der Grafiken (lattice) wird
ueber \code{grafik = "ggplot2like"} gesteuert #' bei \code{grafik = 1} wird die normale
Einstellung verwendet. Unter \code{options()$stp25}
findet man die Liste mit allen Einstellungen.
Die HTML Ausgave wur ?ber \code{Projekt("html")} gesteuert und durch \code{End()} wird die Seite geladen.
}
\subsection{Tabellen}{
APA -Style Tabellen erstellten (HTML oder Console) \code{APA2}.
}
\subsection{Text}{
APA -Style Daten ausgabe als Text kann mit (HTML oder Console)
 \code{Result} erstellten.
}
\subsection{Daten}{
 Unterschiedliche Daten-Formate einlesen (SPss, csv usw) mit
 \code{GetData}. Schreiben lassen sich die
 Daten (auch Grafiken) mit \code{SaveDate}
 }
\subsection{Messinstrumente}{
Fuer spezielle Messinstrumente wie \code{Kano}, \code{Likert},
oder \code{Rangreihe} stehen eigene Funktionen zur verfuegung.
}
\subsection{Aggregieren}{
Hilfsfunktionen zum Aggregieren finden sich unter \code{dapply2},
eine \code{apply()} Variante.
und \code{Melt2} eine Kombination aus \code{melt} und \code{dcast}
aus dem Packet \code{reshape2}.
Melt2 wird dabei wie APA2 verwendet also zB
\code{APA2(Produkt + Online + Preis ~ Group, DF, fun=mean2)}ist
fuer APA-Tabelle und wenn die Mittelwerte angefordert werden
sollen \code{Melt2(Produkt + Online + Preis ~ Group, DF, fun = mean2)}
 die Syntax. Summenindex kann mit \code{Index} oder das Umkodieren
 von Data.frames durch \code{Umcodieren}
erfolgen. Normskalen wie T-Werte lassen sich ueber \code{Norms} berechnen.
}
\subsection{Berechnen}{
Mittelwerte mit \code{mean2} Formatierte Masszahlen fuer
die \code{APA2.formula(~a+d, data, fun=Prozent)} zu finden
unter \code{Prozent}
}

End: Zuruecksetzen der Einstellungen und Aufruf des Browsers browser = getOption("browser")

Methode, Materials, Research_Design, Measures
Results, Demographic_Variables und Statistic sind Platzhalter Funktionen um den R-Code in verschiedenen Files aiszulagern.
}
\examples{
# require(stpvers)
# Projekt()
# APA2(~. , hkarz)
set_my_options(prozent=list(digits=c(1,0), style=2))
get_my_options()$apa.style$prozent

set_my_options(mittelwert=list(digits=c(1,0), plusmin_sign=TRUE))
get_my_options()$apa.style$mittelwert

# APA2(~. , hkarz)

# End()
}
