% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{Cs}
\alias{reexports}
\alias{contr.Treatment}
\alias{reexports}
\alias{contr.Sum}
\alias{reexports}
\alias{contr.Helmert}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{car}{\code{\link[car]{contr.Treatment}}, \code{\link[car]{contr.Sum}}, \code{\link[car]{contr.Helmert}}}

  \item{Hmisc}{\code{\link[Hmisc]{Cs}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

