% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveData.R
\name{SaveData}
\alias{SaveData}
\title{Speichern von Grafiken}
\usage{
SaveData(caption = "", filename = "", w = dev.size("in")[1],
  h = dev.size("in")[2], save_plot = TRUE, output = which_output())
}
\arguments{
\item{caption}{Grafik Abbildung  caption=data,}

\item{filename}{Daten oder Character fuer die speicherung einer Grafik}

\item{save_plot}{speichern als file = TRUE,}

\item{Width, Height, w, h}{an HTMLplot default = 520, Breite der  Grafik  also zb w=8=dev.size("in")[2], h=dev.size("in")[1] Hoehe der Grafik also zb h=8}

\item{GraphBorder}{Rand um Grafik = 0}

\item{Align}{Centriert}

\item{GraphBackGround}{Hintergrund= "white"}

\item{...}{zusaetliche Parameter}
}
\description{
Grafik werden  als PDF oder Windows-Metafile
speichern. Die Funktion Abb() beinhaltet den Pfad.
Die Funktion arbeitet ueber  R2HTML::HTMLplot()
}
