% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lattice.R
\name{set_lattice}
\alias{set_lattice}
\alias{set_lattice_ggplot}
\alias{set_lattice_bw}
\alias{reset_lattice}
\title{set_lattice}
\usage{
set_lattice_ggplot(
  pch = 15:18,
  lty = 1:3,
  cex = 1,
  col = NULL,
  col.bar = NULL,
  strip.background.col = c("grey80", "grey70", "grey60"),
  ...
)

set_lattice_bw(
  pch = 15:18,
  lty = 1:3,
  cex = 1,
  col = grey.colors(7, start = 0.3, end = 0.9),
  col.bar = "grey50",
  strip.background.col = "white",
  ...
)

set_lattice(theme = NULL)

reset_lattice(...)
}
\arguments{
\item{pch, lty, cex}{parameter an plot}

\item{col}{Farbe 
  rosa-himmelblau brewer.pal(8,"Set3")[c(3,4)] 
Grau   grey.colors(5, start = 0.3, end = 0.9) oder gray(seq(0,.9,len = 25))}

\item{col.bar}{Farben}

\item{strip.background.col}{set_lattice_bw: "white" sonst  c("grey80", "grey70", "grey60"),}

\item{...}{nicht benutzt}

\item{theme}{ggplot2like(), custom.theme(), theEconomist.theme}
}
\value{
im Hintergrund oopt unt opar sowie  invisible(lattice::trellis.par.get())
}
\description{
set_lattice_ggplot:   lattice-Theme

set_lattice_bw:  schwarz-weis lattice-Theme

set_lattice: set and reset of lattice default options
}
\examples{
#' 
set.seed(2)
n <- 20 * 3 * 2
DF <- data.frame(
  n = runif(n, min = 1, max = 5),
  e = runif(n, min = 1, max = 5),
  o = runif(n, min = 1, max = 5),
  g = runif(n, min = 1, max = 5),
  a = runif(n, min = 1, max = 5),
  treatment = gl(3, n / 3, labels = c("UG1", "UG2", "KG"))[sample.int(n)],
  sex = gl(2, n / 2, labels = c("male", "female"))
)



set_lattice_ggplot()
bwplot2(e ~ treatment,
        DF,
        groups = sex,
        auto.key = list(columns = 2))


set_lattice_bw()
bwplot2(e ~ treatment,
        DF,
        groups = sex,
        auto.key = list(columns = 2))
 
}
