% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settings}
\alias{settings}
\alias{set_my_options}
\alias{get_my_options}
\alias{set_default_params}
\alias{stp25_options}
\alias{default_stp25_opt}
\title{Einstellungen}
\usage{
set_my_options(...)

get_my_options(x = "stp25")

set_default_params(params)

stp25_options()

default_stp25_opt(
  myformat = "",
  fig_folder = "Fig",
  graphics_devices = "pdf",
  temp_dir = tempdir(),
  data_folder = "Raw data",
  html_folder = "Results",
  OutDec = ".",
  Align = "center",
  language = "de",
  digits.prozent = 0,
  digits.mean = 2,
  digits.r = 2,
  digits.p = 3,
  digits.Fstat = 2,
  null_percent_sign = ".",
  lead.zero = FALSE,
  p.values = c(0.001, 0.01, 0.05, 0.1)[1:3],
  signif.stars = FALSE,
  plusmin_sign = FALSE,
  sep_element = ",",
  brackets = c("[", "]"),
  digits_type = "digits",
  prozent_style = 1,
  median.style = "IQR"
)
}
\arguments{
\item{...}{zu aendernde Einstellungen}

\item{x}{Library = "stp25"}

\item{params}{List of parameters.}

\item{myformat}{HTML, spin, oder Text  spin ist knitr wobei die Ausgabe der Tabellen mit html erfolgt}

\item{fig_folder, temp_dir, data_folder, html_folder}{Folder Fig, tempdir(), Raw data, Results}

\item{graphics_devices}{Grafik als PDF speichern}

\item{OutDec}{Dezimaltrennzeichen}

\item{Align}{Zentrieren}

\item{language}{Sprache}

\item{digits.prozent, digits.mean, digits.r, digits.p, digits.Fstat}{Digits fuer Format2}

\item{null_percent_sign, lead.zero, p.values, signif.stars, plusmin_sign, sep_element, brackets}{Diverse Formatierungen}

\item{digits_type, prozent_style, median.style}{digits_type,  #- entwerde Runden auf zwei stellen oder auf Signifikante Stellen prozent_style 2=medizin Anzahl(Prozent)}
}
\description{
Einstellungen stp25, die Einstellungen betreffen vor allem die Formatierung von Zahlen in den Tabellen.

set_my_options:  Optionen erstellen

get_my_options: Abruf der Optionen

set_default_params: Speichern von Variablen im Global Envirment
Set default parameters: Stolen from library(ezknitr).

stp25_options: Speichern die Default-Optionen in options$stp25

default_stp25_opt: Liste mit den Optionen
}
\examples{
options("stp25" = default_stp25_opt(language = "en"))

#-- default options
set_my_options()
get_my_options()$apa.style$prozent

#-- Speicherort aendern
get_my_options()$fig_folder
#set_my_options(fig_folder="Fig2/")
get_my_options()$fig_folder

#-- Format aendern
#set_my_options(prozent=list(digits=c(1,0), style=2))

#set_my_options(mittelwert=list(digits=c(1,0), plusmin_sign=TRUE))
#set_my_options(prozent = list(null_percent_sign = " . "))

#  APA2(~. , hkarz)
##  #Names2Language(c("Item", "Characteristics", "Statistics"  ), "de")


# options()$stp25$bez$f.value


# Impliziet (also nicht als default vorhanen)
#set_my_options(mittelwert=list(include_name=FALSE)))


}
