% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stp25output.R
\docType{package}
\name{stp25output-package}
\alias{stp25output}
\alias{stp25output-package}
\title{Generate R documentation from inline comments.}
\description{
Roxygen2 allows you to write documentation in comment blocks co-located
with code.
}
\details{
The only function you're likely to need from \pkg{roxygen2} is
\code{\link{roxygenize}}. Otherwise refer to the vignettes to see
how to format the documentation.
}
