% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_abb.R
\name{Tab}
\alias{Tab}
\title{Tabellen-Nummern}
\usage{
Tab(text = "")
}
\arguments{
\item{text}{Name der Tabelle}
}
\value{
Name plus Nummer
}
\description{
Nummerierte Tabellen, mit aufruf der Funktion \code{Abb()} eird eine Tabellen Nummerierung
fortlaufend erstellt. Die Funktion wird intern in der \code{APA2(...)} verwendet.
Die Nummer wird dabei in der Veriable \code{Tab_Index} hinterlegt.
}
\author{
Wolfgang Peter
}
