% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_abb.R
\name{Tab_Abb}
\alias{Tab_Abb}
\alias{Tab}
\alias{Abb}
\alias{Caption}
\alias{Note}
\title{Tabellen und Abbildungs Nummern}
\usage{
Tab(caption = "")

Abb(filename = "", caption = "", folder = options()$stp25$fig_folder)

Caption(caption = NULL, atr = NULL, output = which_output())

Note(x = NULL, atr = NULL)
}
\arguments{
\item{caption}{}

\item{filename}{Name des Foles}

\item{folder}{Speicherort}

\item{atr}{in Caption: alternativer Text}

\item{x}{Text}
}
\value{
Name plus Nummer
}
\description{
Nummerierte Abbildung default Folder ist fig

Ueberschrift aus stp-Objekt: Caption(caption, attr(x, "caption"))
oder Note(note, attr(x, "note")) mit
}
