% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.data.frame.R
\name{tbl_header}
\alias{tbl_header}
\title{Header aufbereiten}
\usage{
tbl_header(x, fix_colnames = FALSE, col.names = colnames(x))
}
\arguments{
\item{x}{data.frame}

\item{fix_colnames}{an translate TRUE/FALSE}

\item{col.names}{colnames(x)}
}
\value{
list(header,header_above,cgroup,n.cgroup,header_above2,)
}
\description{
Interne Funktion
}
