% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.R
\name{Text}
\alias{Text}
\alias{Head}
\alias{Anmerkung}
\alias{CitationLib}
\alias{Zitat}
\alias{Arbeitszeit}
\alias{Anhang}
\title{Text Ausgabe}
\usage{
Text(..., style = 0, char = "-", output = which_output())

Head(..., style = 3, char = "-")

Anmerkung(...)

CitationLib(lib)

Zitat(...)

Arbeitszeit(Lines, sep = "\\\\t")

Anhang(include.literatur = TRUE, include.software = TRUE,
  include.mean = TRUE, include.test = APA_formula,
  include.regression = FALSE, include.lm = APA_regression,
  include.binom = APA_regression, include.lme = APA_regression,
  include.poisso = APA_regression, include.rangreihe = APA_regression,
  include.aufbau = TRUE, include.abkuerzung = TRUE, ...)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}

\item{style}{0 = p oder normaler Text 2 = h2 also Ueberschrift bei Consolen-Ausgabe}

\item{char}{ist nicht zu Verwenden Text Trennzeichen bei Consolen-Ausgabe}

\item{output}{Abrage nach dem Ausgabeformat  which_output()}

\item{lib}{Paketname}

\item{Lines}{in Arbeitszeit der Input-String}

\item{sep}{in Arbeitszeit das Input-String-Trennzeichen}

\item{include.literatur}{Literatur}

\item{include.software}{Verwendete Software}

\item{include.mean}{Mittelwerte Prozent}

\item{include.test}{Signifikanz Test}

\item{include.regression}{Regression Allgenein}

\item{include.lm}{Regression linear}

\item{include.binom}{Regression binomial}

\item{include.lme}{Regression Hirarchisch}

\item{include.poisso}{Regression Poisson}

\item{include.rangreihe}{Rangreihe}

\item{include.aufbau}{Uberschriften Aufbau}

\item{include.abkuerzung}{Abkuerzungen}
}
\value{
Vector
}
\description{
Ausgabe von \code{Text()} und Ueberschriften. \code{Head} ist dabei eine Kopie von \code{Text} mit
dem Parameter \code{style = 2}#'

Anmerkung() ist ein blauer Text.

CitationLib(car) gibt Pakete beschreibungen aus.

Zitat() ist eine Text  vom Type <blockquote>

Arbeitszeit() Tabelle zur Dokumentation der Arbeiszeit.
}
\examples{
#library(stp25)
#Projekt("html")
Head("Ueberschrift") ## HR=2

Text("hallo", "ich bin ein Test")
###plot(sin, -pi, 2*pi,main="Sinus")
###HTMLplot( Caption="Look at this curve!")
#End()
}
