% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R, R/Output.data.frame.R,
%   R/Output_kable.R, R/apa_table.R, R/xtabs.R
\name{Output}
\alias{Output}
\alias{Output.character}
\alias{Output.vector}
\alias{Output.htmlTable}
\alias{Output.NULL}
\alias{Output.default}
\alias{Output.list}
\alias{Output.matrix}
\alias{Output.stp25}
\alias{Output_kable}
\alias{Output_kable.list}
\alias{Output_kable.default}
\alias{Output.apa_table}
\alias{Output.table}
\alias{Output.xtabs}
\alias{Output.ftable}
\title{Ausgabe von HTML, Knit oder Text}
\usage{
Output(x, ...)

\method{Output}{character}(x, output = which_output(), ...)

\method{Output}{vector}(x, output = which_output(), ...)

\method{Output}{htmlTable}(x, output = which_output(), ...)

\method{Output}{NULL}(x, ...)

\method{Output}{default}(x, ...)

\method{Output}{list}(x, ...)

\method{Output}{matrix}(x, ...)

\method{Output}{stp25}(x, ...)

Output_kable(x, ...)

\method{Output_kable}{list}(x, ...)

\method{Output_kable}{default}(x, caption = NULL,
  col.names = colnames(x), booktabs = TRUE,
  latex_options = c("hold_position"), note = NULL,
  output = which_output(), print_col = NULL, fix_colnames = TRUE,
  ...)

\method{Output}{apa_table}(x)

\method{Output}{table}(x, caption = attr(x, "caption"), note = attr(x,
  "note"), ...)

\method{Output}{xtabs}(x, caption = attr(x, "caption"), note = attr(x,
  "note"), ...)

\method{Output}{ftable}(x, caption = attr(x, "caption"), note = attr(x,
  "note"), ...)
}
\arguments{
\item{x}{Objekt liste oder Dataframe}

\item{...}{weitere Einstellungen}

\item{output}{HTML, Knit  oder Textfile output=NA verlaesst die Funktion ohne aktion}

\item{col.names}{Output_kable:  fuer tintPdf format = "latex"}

\item{booktabs}{kable an Latex Latex}
}
\value{
HTML oder Textausgabe
}
\description{
Erstellt den Output von Objekten, diese koennen zB. mit Uberschriften versehen sein.

Output.htmlTable: Verarbeitet htmlTable-Objekt als auch Listen
 oder auch einfach nur ein string. Die Tabelle mit Header usw ist dabei
 schon vorbereitet. Bei Markdown kommt output=FALSE

Output.iste: einzeln in einen data.frame transformieren

Output.matrix: umwandeln in einen data.frame

Output.stp25: experimenteller Prototyp

Output mit  knitr::kable Funftioniert nur wenn options(knitr.table.format = "latex")
gesetzt wird (das gilt nur fuer Pdf)

Output fur apaTables-Objekte wie zB apa.reg.table

Output.xtabs und Output.tabl und Output.ftable fuer Tabellen Objekte
}
\examples{

\dontrun{
library(apaTables) 
 basic.reg <- lm(sales ~ adverts + airplay, data=album)
apa.reg.table(basic.reg) \%>\% Output()


lm_output <- lm(libido ~ dose, data=viagra)
apa.aov.table(lm_output) \%>\% Output()

APA_Table(lm_output)

APA_Table( aov(libido ~ dose, data=viagra))

block1 <- lm(sales ~ adverts + airplay, data=album)
block2 <- lm(sales ~ adverts*airplay, data=album)
apa.reg.table(block1, block2) \%>\% Output()
 } 

}
\author{
Wolfgang Peter
}
