% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R, R/Output.apa_table.R,
%   R/Output.data.frame.R, R/Output_kable.R, R/Text.R, R/xtabs.R
\name{Output}
\alias{Output}
\alias{Output.xtable}
\alias{Output.character}
\alias{Output.vector}
\alias{Output.htmlTable}
\alias{Output.NULL}
\alias{Output.default}
\alias{Output.psychobject}
\alias{Output.emmGrid}
\alias{Output.apa_table}
\alias{Output.list}
\alias{Output.matrix}
\alias{Output.stp25}
\alias{add_row_df}
\alias{Output_kable}
\alias{Output_kable.list}
\alias{Output_kable.default}
\alias{Text.default}
\alias{Text.character}
\alias{Output.table}
\alias{Output.xtabs}
\alias{Output.ftable}
\title{Ausgabe von HTML, Knit oder Text}
\usage{
Output(x, ...)

\method{Output}{xtable}(
  x,
  caption = attr(x, "heading"),
  digits = attr(x, "digits"),
  output = which_output(),
  ...
)

\method{Output}{character}(x, output = which_output(), ...)

\method{Output}{vector}(x, output = which_output(), ...)

\method{Output}{htmlTable}(x, output = which_output(), ...)

\method{Output}{`NULL`}(x, ...)

\method{Output}{default}(x, ...)

\method{Output}{psychobject}(x, digits = c(2, 2, 1, 3, 2, 2), ...)

\method{Output}{emmGrid}(
  x,
  caption = "Estimated marginal means",
  note = "Least-squares means",
  ...
)

\method{Output}{apa_table}(x, output = which_output())

\method{Output}{list}(x, output = which_output(), ...)

\method{Output}{matrix}(x, output = which_output(), ...)

\method{Output}{stp25}(x, output = which_output(), ...)

add_row_df(x, add_row = NULL)

Output_kable(x, ...)

\method{Output_kable}{list}(x, ...)

\method{Output_kable}{default}(
  x,
  caption = NULL,
  col.names = colnames(x),
  booktabs = TRUE,
  latex_options = c("hold_position"),
  note = NULL,
  output = which_output(),
  print_col = NULL,
  fix_colnames = options()$stp25$language != "",
  ...
)

\method{Text}{default}(x, ...)

\method{Text}{character}(..., style = 0, char = "-", output = which_output())

\method{Output}{table}(x, caption = attr(x, "caption"), note = attr(x, "note"), ...)

\method{Output}{xtabs}(x, caption = attr(x, "caption"), note = attr(x, "note"), ...)

\method{Output}{ftable}(x, caption = attr(x, "caption"), note = attr(x, "note"), ...)
}
\arguments{
\item{x}{Objekt liste oder Dataframe}

\item{...}{weitere Einstellungen}

\item{output}{FALSE = NULL,
TRUE = print,
"html" = htmlTable(), "markdown_html"= htmlTable(),
"markdown"= kable(),
"text"=kable(... "pandoc") und
"word"= flextable()}

\item{add_row}{list (c("Erste Zeile" = 1, "Dritte" = 3))}

\item{col.names}{Output_kable:  fuer tintPdf format = "latex"}

\item{booktabs, latex_options}{kable an Latex     latex_options = c("hold_position"),
latex_options = "scale_down"}
}
\value{
HTML oder Textausgabe
}
\description{
HTML :  htmlTable::htmlTable

Output.xtable mit xtable lassen sich alternativ r-Objekte ausgeben.

Output.htmlTable: Verarbeitet htmlTable-Objekt als auch Listen
 oder auch einfach nur ein string. Die Tabelle mit Header usw ist dabei
 schon vorbereitet. Bei Markdown kommt output=FALSE

Output fur apaTables-Objekte wie zB apa.reg.table

Output.iste: einzeln in einen data.frame transformieren

Output.matrix: umwandeln in einen data.frame

Output.stp25: experimenteller Prototyp

Output mit  knitr::kable Funftioniert nur wenn options(knitr.table.format = "latex")
gesetzt wird (das gilt nur fuer Pdf)

Output.xtabs und Output.tabl und Output.ftable fuer Tabellen Objekte
}
\details{
markdown: knitr::kable -> kableExtra::kable_styling 

Text: knitr::kable

Word: knitr::kable(x, format = "pandoc")  

Output_word() flextable::regulartable
}
\examples{

 require(xtable)
## Load example dataset
data(tli)

## Demonstrate aov
fm1 <- aov(tlimth ~ sex + ethnicty + grade + disadvg, data = tli)
fm1.table <- xtable(fm1)

Output(fm1.table)

# ## Demonstrate lm
fm2 <- lm(tlimth ~ sex*ethnicty, data = tli)
fm2.table <- xtable(fm2)
fm2b <- lm(tlimth ~ ethnicty, data = tli)

Output(fm2.table)
Output(xtable(anova(fm2)))
Output(xtable(anova(fm2b, fm2)))


warp.lm <- lm(breaks ~ wool * tension, data = warpbreaks)
x <- emmeans::emmeans(warp.lm,  ~ wool | tension)
Output(x)


\dontrun{
library(apaTables) 
 basic.reg <- lm(sales ~ adverts + airplay, data=album)
apa.reg.table(basic.reg) \%>\% Output()


lm_output <- lm(libido ~ dose, data=viagra)
apa.aov.table(lm_output) \%>\% Output()

APA_Table(lm_output)

APA_Table( aov(libido ~ dose, data=viagra))

block1 <- lm(sales ~ adverts + airplay, data=album)
block2 <- lm(sales ~ adverts*airplay, data=album)
apa.reg.table(block1, block2) \%>\% Output()
 } 

df <-   data.frame(
Source = c("A", "B", "C", "F"),
x = 1:4,
y = 1:4,
stringsAsFactors = FALSE
)

# add_row_df(df, c("Erste Zeile" = 1, "Dritte" = 3))


}
\author{
Wolfgang Peter
}
