% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projekt.R, R/R2HTML.R
\name{Projekt}
\alias{Projekt}
\alias{Projekt_Rmd}
\alias{End}
\alias{Rechnung}
\alias{Methode}
\alias{Materials}
\alias{Research_Design}
\alias{Measures}
\alias{Results}
\alias{Demographic_Variables}
\alias{Statistic}
\alias{get_scriptpath}
\alias{HTML_Start}
\alias{HTML_End}
\title{Projekt}
\usage{
Projekt(
  myformat = "",
  Projektname = "Demo",
  datum = date(),
  fig_folder = "Fig",
  html_folder = "Results",
  OutDec = NULL,
  contrasts = c("contr.Treatment", "contr.poly"),
  css = TRUE,
  ...
)

Projekt_Rmd(myformat, Projektname, datum, OutDec, contrasts)

End(
  anhang = FALSE,
  browser = "C:/Program Files (x86)/Internet Explorer/iexplore.exe",
  output = options()$prompt[1] == "HTML> ",
  ...
)

Rechnung(datum = "")

Methode(h = "Methoden", x = NULL, file = NULL)

Materials(h = "Materialien", x = NULL, file = "(1) Get Data.R")

Research_Design(h = "Forschungs Design", x = NULL, file = NULL)

Measures(h = "Messinstrument", x = NULL, file = "(2) Measures.R")

Results(h = "Ergebnisse", x = NULL, file = NULL)

Demographic_Variables(
  h = "Deskriptive Analyse",
  x = NULL,
  file = "(3) Demographic.R"
)

Statistic(h = "Resultate", x = NULL, file = "(4) Analyse.R")

get_scriptpath()

HTML_Start(
  Projektname = "Demo",
  datum = date(),
  fig_folder = "Fig",
  html_folder = "Results",
  OutDec = NULL,
  contrasts = c("contr.Treatment", "contr.poly"),
  withprompt = "HTML> ",
  .extension = "html",
  .myDir = getwd(),
  .filename = cleansing_umlaute(Projektname),
  .output.dir = file.path(.myDir, html_folder),
  .HTML.file = file.path(.output.dir, paste(.filename, ".", .extension, sep = "")),
  .css.file = file.path(.output.dir, "layout.css")
)

HTML_End(browser = "iexplore")
}
\arguments{
\item{myformat}{HTML, Spin, Knit, Rpres oder Text. Spin ist knitr wobei die Ausgabe der Tabellen mit html erfolgt}

\item{Projektname}{Bezeichnung des Projektes (gilt auch fuer die HTML Seite)}

\item{datum}{Datum zur Dokumentation}

\item{fig_folder, html_folder}{Folder bei html wenn ein ander Ort gewuenscht}

\item{OutDec}{Komma oder Punkt}

\item{contrasts}{default wie SPSS}

\item{css}{css Eigense Format}

\item{...}{weitere Objekte nicht benutzt}

\item{anhang}{Ja/Nein}

\item{browser}{Ie oder Chrome}

\item{output}{TRUE oder FALSE}

\item{h}{Ueberschrift}

\item{x}{Character}

\item{file}{auszufuerendes File}
}
\value{
html-Pfad oder Projektname als Text
}
\description{
Ein R-File ruft die Interne Funktion Projekt_html() auf und ein Rmd-File Projekt_Rmd().

Einstellungen fuer .Rmd files hier werden keine Folder erstellt

\subsection{End}{
Zuruecksetzen der Einstellungen und Aufruf des Browsers browser = getOption("browser")}

Rechnung(): setzt ein Datum und Beendet mit Ende und stop die Auswertung

\subsection{Methode}{
Methode, Materials, Research_Design, Measures
Results, Demographic_Variables und Statistic sind Platzhalter Funktionen um den R-Code
in verschiedenen Files aiszulagern.}

Results(), Measures(), Materials(), usw Ueberschrift mit aufruf eies exteren R-Scripts.

\subsection{Interne Funktion}{
\code{get_scriptpath()} Ausfuehrendes File finden
Quelle: https://stackoverflow.com/questions/18000708/find-location-of-current-r-file}

HTML_Start ist ein e Kopie von  R2HTML
}
\details{
Start eines neuen Prpjekts, mit \code{Projekt()} werden die default Einstellungen (Kontraste) gesetzt sowie die Folder erstellt.
Weiters wird die Ausgabe als HTML gesteuert.
Weitere Optionen fuer die Formatierung koennen mit \code{default_stp25_opt()} gesetzt.
Mit  \code{stp25_options} koennen die  Formatierungs-Optionen geaendert werden.
}
\examples{

\dontrun{
# require(stpvers)

 Projekt()
     # R2HTML::HTMLChangeCss("gridR2HTML")
     # options(prompt="HTML> ")
     #  set_default_params(list(myURL = myURL))

 APA2(~. , hkarz)


#-- default options
set_my_options()
get_my_options()$apa.style$prozent

#-- Speicherort aendern
get_my_options()$fig_folder
set_my_options(fig_folder="Fig2/")
get_my_options()$fig_folder

#-- Format aendern
set_my_options(prozent=list(digits=c(1,0), style=2))

set_my_options(mittelwert=list(digits=c(1,0), plusmin_sign=TRUE))


#  APA2(~. , hkarz)
###Names2Language(c("Item", "Characteristics", "Statistics"  ), "de")


# options()$stp25$bez$f.value
 ### Names2Language(c("Pr..Chisq.", "F.value"))


#  hkarz2 <-Label(mutate(hkarz,
#  Lai=factor(lai),
#  x=NA),
#  Lai="1-Smoking Habits",
#  x =   "2-Supplement Intake",
#  tzell = " - to boost your performance" ) 

#  Tabelle2(  ~Lai+x[header]+tzell,  hkarz2  , APA=TRUE )
#  Tabelle2(  Lai+x[header]+tzell~gruppe, hkarz2 , test=TRUE, APA=TRUE, include.total=TRUE)
#  set_my_options(mittelwert=list(include_name=FALSE))
#  Tabelle2(  Lai+x[header]+tzell[median]~gruppe, hkarz2 , test=TRUE, APA=TRUE, include.total=TRUE)

#APA2(~. , hkarz)

 End()

}

}
