% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projekt.R
\name{Projekt}
\alias{Projekt}
\alias{End}
\alias{Methode}
\alias{Materials}
\alias{Research_Design}
\alias{Measures}
\alias{Results}
\alias{Demographic_Variables}
\alias{Statistic}
\alias{get_scriptpath}
\title{Start eines neuen Prpjekts}
\usage{
Projekt(myformat = "", Projektname = "Demo", datum = date(),
  fig_folder = "Fig", html_folder = "Results", OutDec = NULL,
  contrasts = c("contr.Treatment", "contr.poly"),
  html_name = stpvers::Clean_Umlaute2(Projektname), css = TRUE, ...)

End(anhang = FALSE,
  browser = "C:/Program Files (x86)/Internet Explorer/iexplore.exe",
  output = options()$prompt[1] == "HTML> ", ...)

Methode(h = "Methoden", x = NULL, file = NULL)

Materials(h = "Materialien", x = NULL, file = "(1) Get Data.R")

Research_Design(h = "Forschungs Design", x = NULL, file = NULL)

Measures(h = "Messinstrument", x = NULL, file = "(2) Measures.R")

Results(h = "Ergebnisse", x = NULL, file = NULL)

Demographic_Variables(h = "Deskriptive Analyse", x = NULL,
  file = "(3) Demographic.R")

Statistic(h = "Resultate", x = NULL, file = "(4) Analyse.R")

get_scriptpath()
}
\arguments{
\item{myformat}{HTML oder Knit oder Console Rpres (R-Presentation)}

\item{Projektname}{Bezeichnung des Projektes (gilt auch fuer die HTML Seite)}

\item{datum}{Datum zur Dokumentation}

\item{fig_folder, html_folder}{Folder wenn ein ander Ort gewuenscht}

\item{OutDec}{Komma oder Punkt}

\item{contrasts}{default wie SPSS}

\item{html_name}{intern nicht aendern}

\item{css}{Eigense Format}

\item{...}{weitere Objekte nicht benutzt}

\item{anhang}{Ja/Nein}

\item{browser}{Ie oder Chrome}

\item{output}{TRUE oder FALSE}

\item{h}{Ueberschrift}

\item{x}{Character}

\item{file}{auszufuerendes File}
}
\value{
Pfad als Text
}
\description{
Mit \code{Projekt()} werden die default Einstellungen (Kontraste) gesetzt sowie die Folder erstellt.
Weiters wird die Ausgabe als HTML gesteuert.
Weitere Optionen fuer die Formatierung koennen mit \code{default_stp25_opt()} gesetzt.
Mit  \code{stp25_options} koennen die  Formatierungs-Optionen geaendert werden.

\subsection{End}{
Zuruecksetzen der Einstellungen und Aufruf des Browsers browser = getOption("browser")}

\subsection{Methode}{
Methode, Materials, Research_Design, Measures
Results, Demographic_Variables und Statistic sind Platzhalter Funktionen um den R-Code
in verschiedenen Files aiszulagern.}

\subsection{Interne Funktion}{
\code{get_scriptpath()} Ausfuehrendes File finden
Quelle: https://stackoverflow.com/questions/18000708/find-location-of-current-r-file}
}
\examples{

\dontrun{
# require(stpvers)

 Projekt()
     # R2HTML::HTMLChangeCss("gridR2HTML")
     # options(prompt="HTML> ")
     #  set_default_params(list(myURL = myURL))

 APA2(~. , hkarz)


set_my_options(prozent=list(digits=c(1,0), style=2))
get_my_options()$apa.style$prozent

set_my_options(mittelwert=list(digits=c(1,0), plusmin_sign=TRUE))
get_my_options()$apa.style$mittelwert

#APA2(~. , hkarz)

 End()

}

}
