% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveData.R
\name{SaveData}
\alias{SaveData}
\title{Speichern von Grafiken}
\usage{
SaveData(
  caption = "",
  w = dev.size("in")[1],
  h = dev.size("in")[2],
  filename = "",
  save_plot = TRUE,
  output = which_output()
)
}
\arguments{
\item{caption}{Grafik Abbildung  caption=data,}

\item{w, h}{Width, Height, an HTMLplot default = 520, Breite der  Grafik  also zb w=8=dev.size("in")[2], h=dev.size("in")[1] Hoehe der Grafik also zb h=8}

\item{filename}{Daten oder Character fuer die speicherung einer Grafik}

\item{save_plot}{speichern als file = TRUE,}
}
\description{
Grafik werden  als PDF oder Windows-Metafile
speichern. Die Funktion Abb() beinhaltet den Pfad.
Die Funktion arbeitet ueber  R2HTML::HTMLplot()
}
