% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.data.frame.R
\name{tbl_header}
\alias{tbl_header}
\title{Header aufbereiten}
\usage{
tbl_header(x, col.names = colnames(x), col_names = NULL,
  fix_colnames = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{col.names}{colnames(x)}

\item{col_names}{optionale Namen}

\item{fix_colnames}{an translate TRUE/FALSE}
}
\value{
list()
}
\description{
Interne Funktion
}
